Extract hashing logic
